+++ CHANGES.md
@@ -1,13 +1,15 @@
 # Changes
 This file marks the changes in each version of ADM starting with 0.2.0.
 
-## 0.4
-### 0.4.2
+## 0.5
+### 0.5.0
 ### Additions
 - Add in --untagged option to the clean command to only remove untagged images (--images --untagged).
+- Add in list command to list all the container names for all or a single application.
 #### Fixes
 - Fix issue when building multiple components that it wouldn't fetch the latest version.
 
+## 0.4
 ### 0.4.1
 #### Additions
 - Add in --force option to the clean command.